Refactor badge update so it's reflected immediately after muting
