+++ src/background/check-pull-requests.ts
@@ -1,7 +1,5 @@
-import { PullRequest } from "../state/storage/last-check";
+import { showBadgeError } from "../badge";
 import { store } from "../state/store";
-import { showBadgeError, updateBadge } from "./badge";
-import { showNotification } from "./notifications";
 
 /**
  * Checks if there are any new pull requests and notifies the user when required.
@@ -14,31 +12,10 @@ export async function checkPullRequests() {
       return;
     }
     await store.github.refreshPullRequests();
-    const unreviewedPullRequests = store.github.unreviewedPullRequests;
-    if (!unreviewedPullRequests) {
-      throw new Error(`Pull requests should have been loaded.`);
-    }
-    await updateBadge(unreviewedPullRequests.length);
-    await showNotificationForNewPullRequests(unreviewedPullRequests);
     error = null;
   } catch (e) {
     error = e;
     await showBadgeError();
   }
   store.github.setError(error ? error.message : null);
-}
-
-/**
- * Shows a notification for each pull request that we haven't yet notified about.
- */
-async function showNotificationForNewPullRequests(pullRequests: PullRequest[]) {
-  for (const pullRequest of pullRequests) {
-    if (!store.github.notifiedPullRequestUrls.has(pullRequest.htmlUrl)) {
-      console.log(`Showing ${pullRequest.htmlUrl}`);
-      showNotification(pullRequest);
-    } else {
-      console.log(`Filtering ${pullRequest.htmlUrl}`);
-    }
-  }
-  store.github.setNotifiedPullRequests(pullRequests);
 }
